# src/test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: support-management
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "12.4"
servers:
  - url: http://localhost:51578
    description: Generated server url
...
paths:
...
  /{municipalityId}/message-exchange-sync-config/{id}:
    put:
      tags:
        - Configuration message-exchange sync
      summary: Update message-exchange sync config
      description: Update message-exchange sync config
      operationId: updateMessageExchangeSyncConfig
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Config id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageExchangeSync"
        required: true
      responses:
        "204":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
        - Configuration message-exchange sync
      summary: Delete message-exchange sync config
      description: Deletes the message-exchange sync config
      operationId: deleteMessageExchangeSyncConfig
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Config id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "204":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/message-exchange-sync-config:
    get:
      tags:
        - Configuration message-exchange sync
      summary: Read message-exchange sync configs
      description: Fetches message-exchange config for a municipality
      operationId: getAllMessageExchangeByMunicipalityId
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageExchangeSync"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
        - Configuration message-exchange sync
      summary: Create message-exchange sync config
      description: Create message-exchange sync config that conversations from message-exchange
        service
      operationId: createMessageExchangeSyncConfig
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageExchangeSync"
        required: true
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: { }
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    MessageExchangeSync:
      type: object
      description: Message-Exchange sync configuration
      properties:
        id:
          type: integer
          format: int64
          description: Unique id
          examples:
            - "1"
          readOnly: true
        namespace:
          type: string
          description: Message exchange namespace to search in. Does not map to supporManagement
            namespace.
          examples:
            - support
        latestSyncedSequenceNumber:
          type: integer
          format: int64
          description: Latest synced sequence number
          examples:
            - "333"
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Timestamp when the configuration was last modified
          examples:
            - 2024-12-24T01:30:00.000+02:00
          readOnly: true
        active:
          type: boolean
          description: If set to true conversations will be synced
      required:
        - active
...
  securitySchemes: { }